requirements.txt
pandas~=1.5.3
matplotlib~=3.6.3
numpy~=1.24.2
Pillow~=9.4.0